temporary frontend modifs

+++ hardhat/contracts/VRFCoordinatorV2Mock.sol
@@ -11,7 +11,7 @@ contract VRFCoordinatorV2Mock {
 
     function requestRandomWords(
         bytes32 keyHash,
-        uint64 subId,
+        uint256 subId,
         uint16 minimumRequestConfirmations,
         uint32 callbackGasLimit,
         uint32 numWords
